feat(update): move and refactor Update COSMO Alpaca system files to GitHub action (#21)
If there are breaking changes in the AL-Go helpers that affect/break our update logic, the update can no longer be executed until the code is manually adjusted. This is because the update code is currently stored locally in each repository.
To break this dependency, we need to move our update code to a separate GitHub action so that we can update it independently.

https://github.com/cosmoconsult/Alpaca-Actions/tree/main/UpdateAlpacaSystemFiles

Closed [AB#4344](https://dev.azure.com/cc-ppi/83f75d99-795d-45dc-8543-9fe1918ff7f9/_workitems/edit/4344)